Fix crash: abs(Int16.min) overflow in SilenceDetector
Same bug as WaveformExtractor — abs(-32768) overflows Int16.
SilenceDetector.detect() crashed on Thread 13 when processing
the 163-minute meeting recording.

Fix: use Int32 for peak amplitude calculation in the window
reduce. Also widen the threshold comparison to Int32.

Verified: no other abs(Int16) calls remain in the codebase.

71/71 EditorCore tests passing.

+++ VideoEditor/Packages/EditorCore/Sources/EditorCore/Analysis/SilenceDetector.swift
@@ -74,8 +74,9 @@ public struct SilenceDetector: Sendable {
                 let window = samples[i..<windowEnd]
 
                 // Calculate peak amplitude in window
-                let peak = window.reduce(Int16(0)) { max(abs($0), abs($1)) }
-                let isSilent = peak < thresholdInt16
+                // Use Int32 to avoid overflow on Int16.min (-32768)
+                let peak = window.reduce(Int32(0)) { max(abs(Int32($0)), abs(Int32($1))) }
+                let isSilent = peak < Int32(thresholdInt16)
 
                 let windowTime = Double(sampleIndex + i) / sampleRate
